refactor(spf): sort features/ into behaviors, actors, primitives
First step in the src/ reorganization ahead of the HLS engine doc.
Kind-separation only — no code changes, no abstraction changes, no
cross-concern moves. Purpose is to make the piles visible so the
later seam-extraction work (substrate, transport, engines) has
something obvious to pull from.

What moved:

src/dom/features/ → split into:
  - src/dom/behaviors/   — the {state, owners[, config]} → cleanup
                           composition integrations (end-of-stream,
                           load-segments, load-text-track-cues,
                           setup-mediasource, setup-sourcebuffer,
                           sync-text-tracks, track-current-time,
                           track-playback-initiated,
                           track-playback-rate, update-duration)
  - src/dom/actors/      — createXActor factories returning message/
                           transition actors (segment-loader,
                           text-track-segment-loader, text-tracks)

src/media/features/ → split into:
  - src/media/behaviors/  — DOM-free composition integrations
                            (calculate-presentation-duration,
                            quality-switching, resolve-presentation,
                            resolve-track, sync-preload-attribute)
  - src/media/primitives/ — pure selector functions (select-tracks)

The -actor suffix drops from filenames now that the actors/ directory
carries the kind. Export names keep the suffix
(createSegmentLoaderActor, TextTracksActor, etc.) so call sites
remain self-describing.

No intentional behavior changes. Build, tests, typecheck, lint,
workspace consistency — all green.
