+++ packages/spf/src/dom/behaviors/end-of-stream.ts
