+++ packages/spf/src/dom/behaviors/sync-text-tracks.ts
