+++ packages/spf/src/dom/behaviors/tests/end-of-stream.test.ts
