+++ packages/spf/src/dom/behaviors/tests/update-duration.test.ts
