+++ packages/spf/src/dom/behaviors/track-current-time.ts
