+++ packages/spf/src/dom/behaviors/update-duration.ts
