+++ packages/spf/src/dom/index.ts
@@ -2,14 +2,14 @@
  * DOM/Browser-specific SPF bindings
  */
 
-export { loadSegments } from './features/load-segments';
-export { loadTextTrackCues } from './features/load-text-track-cues';
-export type { CurrentTimeOwners, CurrentTimeState } from './features/track-current-time';
-export { trackCurrentTime } from './features/track-current-time';
-export type { PlaybackInitiatedOwners, PlaybackInitiatedState } from './features/track-playback-initiated';
-export { trackPlaybackInitiated } from './features/track-playback-initiated';
-export type { PlaybackRateOwners, PlaybackRateState } from './features/track-playback-rate';
-export { trackPlaybackRate } from './features/track-playback-rate';
+export { loadSegments } from './behaviors/load-segments';
+export { loadTextTrackCues } from './behaviors/load-text-track-cues';
+export type { CurrentTimeOwners, CurrentTimeState } from './behaviors/track-current-time';
+export { trackCurrentTime } from './behaviors/track-current-time';
+export type { PlaybackInitiatedOwners, PlaybackInitiatedState } from './behaviors/track-playback-initiated';
+export { trackPlaybackInitiated } from './behaviors/track-playback-initiated';
+export type { PlaybackRateOwners, PlaybackRateState } from './behaviors/track-playback-rate';
+export { trackPlaybackRate } from './behaviors/track-playback-rate';
 export { appendSegment } from './media/append-segment';
 export { flushBuffer } from './media/buffer-flusher';
 export type { SpfMediaAPI, SpfMediaProps } from './playback-engine/adapter';
