+++ packages/spf/src/dom/playback-engine/hls-engine.ts
@@ -1,23 +1,23 @@
 import { type Composition, createComposition } from '../../core/composition/create-composition';
 import type { ReadonlySignal, Signal } from '../../core/signals/primitives';
 import type { BandwidthState } from '../../media/abr/bandwidth-estimator';
-import { calculatePresentationDuration } from '../../media/features/calculate-presentation-duration';
-import { switchQuality } from '../../media/features/quality-switching';
-import { resolvePresentation } from '../../media/features/resolve-presentation';
-import { resolveTrack } from '../../media/features/resolve-track';
-import { selectAudioTrack, selectTextTrack, selectVideoTrack } from '../../media/features/select-tracks';
-import { syncPreloadAttribute } from '../../media/features/sync-preload-attribute';
-import { endOfStream } from '../features/end-of-stream';
-import { loadSegments } from '../features/load-segments';
-import { loadTextTrackCues } from '../features/load-text-track-cues';
-import { setupMediaSource } from '../features/setup-mediasource';
-import { setupSourceBuffers } from '../features/setup-sourcebuffer';
-import { syncTextTracks } from '../features/sync-text-tracks';
-import type { TextTrackSegmentLoaderActor } from '../features/text-track-segment-loader-actor';
-import type { TextTracksActor } from '../features/text-tracks-actor';
-import { trackCurrentTime } from '../features/track-current-time';
-import { trackPlaybackInitiated } from '../features/track-playback-initiated';
-import { updateDuration } from '../features/update-duration';
+import { calculatePresentationDuration } from '../../media/behaviors/calculate-presentation-duration';
+import { switchQuality } from '../../media/behaviors/quality-switching';
+import { resolvePresentation } from '../../media/behaviors/resolve-presentation';
+import { resolveTrack } from '../../media/behaviors/resolve-track';
+import { syncPreloadAttribute } from '../../media/behaviors/sync-preload-attribute';
+import { selectAudioTrack, selectTextTrack, selectVideoTrack } from '../../media/primitives/select-tracks';
+import type { TextTrackSegmentLoaderActor } from '../actors/text-track-segment-loader';
+import type { TextTracksActor } from '../actors/text-tracks';
+import { endOfStream } from '../behaviors/end-of-stream';
+import { loadSegments } from '../behaviors/load-segments';
+import { loadTextTrackCues } from '../behaviors/load-text-track-cues';
+import { setupMediaSource } from '../behaviors/setup-mediasource';
+import { setupSourceBuffers } from '../behaviors/setup-sourcebuffer';
+import { syncTextTracks } from '../behaviors/sync-text-tracks';
+import { trackCurrentTime } from '../behaviors/track-current-time';
+import { trackPlaybackInitiated } from '../behaviors/track-playback-initiated';
+import { updateDuration } from '../behaviors/update-duration';
 import type { SourceBufferActor } from '../media/source-buffer-actor';
 import { destroyVttParser } from '../text/parse-vtt-segment';
