+++ packages/spf/src/dom/tests/load-segments-track-switch.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { signal } from '../../core/signals/primitives';
 import type { Presentation, VideoSelectionSet } from '../../media/types';
-import type { SegmentLoadingOwners, SegmentLoadingState } from '../features/load-segments';
-import { loadSegments } from '../features/load-segments';
+import type { SegmentLoadingOwners, SegmentLoadingState } from '../behaviors/load-segments';
+import { loadSegments } from '../behaviors/load-segments';
 import { createSourceBufferActor } from '../media/source-buffer-actor';
 
 // ============================================================================
