+++ packages/spf/src/media/behaviors/calculate-presentation-duration.ts
